refactor: FestivalButton 컴포넌트에서 delete페이지로 체크박스 이동
